refactor(release): keep the shared verifier module to what is actually shared
